% Auto-generated: do not edit by hand
\name{''ChartEditor}

\alias{''ChartEditor}

\title{ChartEditor component}

\description{

}

\usage{
''ChartEditor(children=NULL, annotateOptions=NULL, controlOptions=NULL,
logoSrc=NULL, logoStyle=NULL, menuPanelOrder=NULL,
structureOptions=NULL, styleOptions=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{annotateOptions}{Logical | lists containing elements 'text', 'shapes', 'images'.
those elements have the following types:
  - text (logical; optional)
  - shapes (logical; optional)
  - images (logical; optional). }

\item{controlOptions}{Logical | lists containing elements 'sliders', 'menus'.
those elements have the following types:
  - sliders (logical; optional)
  - menus (logical; optional). }

\item{logoSrc}{Character. }

\item{logoStyle}{Named list. }

\item{menuPanelOrder}{Unnamed list. }

\item{structureOptions}{Logical | lists containing elements 'traces', 'subplots', 'transforms'.
those elements have the following types:
  - traces (logical; optional)
  - subplots (logical; optional)
  - transforms (logical; optional). }

\item{styleOptions}{Logical | lists containing elements 'general', 'traces', 'axes', 'maps', 'legend', 'colorbars'.
those elements have the following types:
  - general (logical; optional)
  - traces (logical; optional)
  - axes (logical; optional)
  - maps (logical; optional)
  - legend (logical; optional)
  - colorbars (logical; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

