% Auto-generated: do not edit by hand
\name{''CustomLogo}

\alias{''CustomLogo}

\title{CustomLogo component}

\description{

}

\usage{
''CustomLogo(src=NULL, style=NULL)
}

\arguments{
\item{src}{Character. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

