% Auto-generated: do not edit by hand
\name{''DashChartEditor}

\alias{''DashChartEditor}

\title{DashChartEditor component}

\description{

}

\usage{
''DashChartEditor(children=NULL, id=NULL, data=NULL, dataSources=NULL,
frames=NULL, layout=NULL, loadFigure=NULL, style=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. }

\item{data}{Logical | numeric | character | named list | unnamed list. }

\item{dataSources}{List with named elements and values of type unnamed list. }

\item{frames}{Logical | numeric | character | named list | unnamed list. }

\item{layout}{Logical | numeric | character | named list | unnamed list. }

\item{loadFigure}{List with named elements and values of type logical | numeric | character | named list | unnamed list. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

